test code
Write a java program to get a positive even integer value from the user and display all possible sum of two odd prime numbers for that number. The integer value should be greater than 10 and less than 50 and it should be an even number. Mention that in the program.
Input Format:

Input consists of an integer value

Output Format:

Refer sample output for formatting specifications

Sample Input:

10

Sample Output:

-> (3, 7) = 10

-> (5, 5) = 10